refactor(shared): extract direct-tool-server helpers into shared.tool_servers
Direct tool server payload handling is common across all tool plugins. Sharing it keeps request-body metadata precedence and system-prompt extraction consistent.
