+++ tools/llm_review.py
@@ -390,6 +390,113 @@ def normalize_terminal_tools_result(
         return terminal_tools
     return {}
 
+# --- inlined from src/owui_ext/shared/tool_servers.py (owui_ext.shared.tool_servers) ---
+import json
+import logging
+from collections.abc import Mapping
+from typing import Any, Optional
+from fastapi import Request
+_tool_servers_log = logging.getLogger("owui_ext.shared.tool_servers")
+
+
+def normalize_direct_tool_servers(value: Any) -> list[dict]:
+    """Normalize direct tool server payload into a list of dict copies."""
+    if not isinstance(value, list):
+        return []
+    normalized = []
+    for item in value:
+        if isinstance(item, dict):
+            normalized.append(dict(item))
+    return normalized
+
+
+def extract_direct_tool_server_prompts(direct_tools: Mapping[str, dict]) -> list[str]:
+    """Collect unique non-empty system prompts from loaded direct tools only."""
+    prompts: list[str] = []
+    seen_prompts: set[str] = set()
+    for tool in direct_tools.values():
+        if not isinstance(tool, dict):
+            continue
+        server = tool.get("server")
+        if not isinstance(server, dict):
+            continue
+        system_prompt = server.get("system_prompt")
+        if isinstance(system_prompt, str):
+            stripped_prompt = system_prompt.strip()
+            if stripped_prompt and stripped_prompt not in seen_prompts:
+                prompts.append(stripped_prompt)
+                seen_prompts.add(stripped_prompt)
+    return prompts
+
+
+async def resolve_direct_tool_servers_from_request_and_metadata(
+    *,
+    request: Optional[Request],
+    metadata: Optional[dict],
+    debug: bool = False,
+) -> list[dict]:
+    """Resolve direct tool servers from request.body() first, then metadata."""
+    metadata_servers = normalize_direct_tool_servers(
+        (metadata or {}).get("tool_servers") if isinstance(metadata, dict) else None
+    )
+    request_servers: list[dict] = []
+    if request is not None:
+        request_body = getattr(request, "body", None)
+        if callable(request_body):
+            try:
+                raw_body = await request_body()
+                if raw_body:
+                    body = json.loads(raw_body)
+                    if isinstance(body, dict):
+                        request_servers = normalize_direct_tool_servers(
+                            body.get("tool_servers")
+                        )
+                        if not request_servers:
+                            nested_metadata = body.get("metadata")
+                            if isinstance(nested_metadata, dict):
+                                request_servers = normalize_direct_tool_servers(
+                                    nested_metadata.get("tool_servers")
+                                )
+            except Exception:
+                request_servers = []
+    if request_servers:
+        if debug and metadata_servers and len(metadata_servers) != len(request_servers):
+            _tool_servers_log.warning(
+                "tool_servers mismatch between request body and metadata; "
+                "using request body tool_servers"
+            )
+        return request_servers
+    return metadata_servers
+
+
+def build_direct_tools_dict(
+    *, tool_servers: list[dict], debug: bool = False
+) -> dict:
+    """Build direct tool entries compatible with Open WebUI middleware."""
+    direct_tools: dict = {}
+    for server in tool_servers:
+        if not isinstance(server, dict):
+            continue
+        specs = server.get("specs", [])
+        if not isinstance(specs, list) or not specs:
+            continue
+        server_payload = {k: v for k, v in server.items() if k != "specs"}
+        for spec in specs:
+            if not isinstance(spec, dict):
+                continue
+            name = spec.get("name")
+            if not isinstance(name, str) or not name:
+                continue
+            direct_tools[name] = {
+                "spec": spec,
+                "direct": True,
+                "server": server_payload,
+                "type": "direct",
+            }
+    if debug and tool_servers and not direct_tools:
+        _tool_servers_log.info("No direct tools loaded from tool_servers")
+    return direct_tools
+
 try:
     import markdown as _markdown_mod
 except ImportError:
@@ -2929,108 +3036,6 @@ async def resolve_terminal_id_from_request_and_metadata(
         return request_terminal_id
 
     return metadata_terminal_id
-
-
-def normalize_direct_tool_servers(value: Any) -> list[dict]:
-    """Normalize direct tool server payload into a list of dict copies."""
-    if not isinstance(value, list):
-        return []
-    normalized = []
-    for item in value:
-        if isinstance(item, dict):
-            normalized.append(dict(item))
-    return normalized
-
-
-def extract_direct_tool_server_prompts(direct_tools: Mapping[str, dict]) -> list[str]:
-    """Collect unique non-empty system prompts from loaded direct tools only."""
-    prompts: list[str] = []
-    seen_prompts = set()
-    for tool in direct_tools.values():
-        if not isinstance(tool, dict):
-            continue
-        server = tool.get("server")
-        if not isinstance(server, dict):
-            continue
-        system_prompt = server.get("system_prompt")
-        if isinstance(system_prompt, str):
-            stripped_prompt = system_prompt.strip()
-            if stripped_prompt and stripped_prompt not in seen_prompts:
-                prompts.append(stripped_prompt)
-                seen_prompts.add(stripped_prompt)
-    return prompts
-
-
-async def resolve_direct_tool_servers_from_request_and_metadata(
-    *,
-    request: Optional[Request],
-    metadata: Optional[dict],
-    debug: bool = False,
-) -> list[dict]:
-    """Resolve direct tool servers from request.body() first, then metadata."""
-    metadata_servers = normalize_direct_tool_servers(
-        (metadata or {}).get("tool_servers") if isinstance(metadata, dict) else None
-    )
-
-    request_servers: list[dict] = []
-    if request is not None:
-        request_body = getattr(request, "body", None)
-        if callable(request_body):
-            try:
-                raw_body = await request_body()
-                if raw_body:
-                    body = json.loads(raw_body)
-                    if isinstance(body, dict):
-                        request_servers = normalize_direct_tool_servers(body.get("tool_servers"))
-                        if not request_servers:
-                            nested_metadata = body.get("metadata")
-                            if isinstance(nested_metadata, dict):
-                                request_servers = normalize_direct_tool_servers(
-                                    nested_metadata.get("tool_servers")
-                                )
-            except Exception:
-                request_servers = []
-
-    if request_servers:
-        if debug and metadata_servers and len(metadata_servers) != len(request_servers):
-            log.warning(
-                "[LLMReview] tool_servers mismatch between request body and metadata; "
-                "using request body tool_servers"
-            )
-        return request_servers
-
-    return metadata_servers
-
-
-def build_direct_tools_dict(*, tool_servers: list[dict], debug: bool) -> dict:
-    """Build direct tool entries compatible with Open WebUI middleware."""
-    direct_tools = {}
-    for server in tool_servers:
-        if not isinstance(server, dict):
-            continue
-
-        specs = server.get("specs", [])
-        if not isinstance(specs, list) or not specs:
-            continue
-
-        server_payload = {k: v for k, v in server.items() if k != "specs"}
-        for spec in specs:
-            if not isinstance(spec, dict):
-                continue
-            name = spec.get("name")
-            if not isinstance(name, str) or not name:
-                continue
-            direct_tools[name] = {
-                "spec": spec,
-                "direct": True,
-                "server": server_payload,
-                "type": "direct",
-            }
-
-    if debug and tool_servers and not direct_tools:
-        log.info("[LLMReview] No direct tools loaded from tool_servers")
-
-    return direct_tools
 
 
 def coerce_user_valves(raw_valves: Any, valves_cls: Type[BaseModel]) -> BaseModel:
